Add lime bg

+++ tools/image.py
@@ -16,7 +16,8 @@ def apply_grey_patch(image, top_left_x, top_left_y, patch_size):
     Returns:
         numpy.ndarray: Patched image
     """
-    patched_image = np.array(image, copy=True)
+    #print("Un de plus !")
+    patched_image = np.array(image, copy=True, dtype='uint8')
     patched_image[
         top_left_y : top_left_y + patch_size, top_left_x : top_left_x + patch_size, :
     ] = 127.5
